Minor improvements
(1)

If we don’t have to ask _match? for a failure, just to then reraise
an exception we should not do that, and instead just pass through
the exception that _match? caught (in the case of a mismatch.) 

(2)

Since the info parameter now has multiple meaning, we move the code
to generate the message into the Error class itself (where it would
have belonged in the first place.)